requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
gspread==5.12.4
google-auth==2.23.4
flask==3.0.0
selenium==4.15.2
webdriver-manager==4.0.1
schedule==1.2.0
python-dotenv==1.0.0
gunicorn==21.2.0
boto3==1.34.0

